Two files are added for classification and regression.
The purpose of an Insurance is to provide protection against the risk of any financial loss. Insurance is a form of risk management in which an insurer agrees to take the risk of the insured entity against future events, uncertain loss due to Tsunami, earthquake or damage against the vehicle or personal property. Here you will be provided with Automobile insurance claim dataset.
One has to predict the claim amount in the Automobile insurance dataset.
One has to predict the insurance fraud  in the Automobile insurance dataset.

We have used regression analysis to check the claim amount the insurance dataset.
We have used classification algorithm to check the fraud detection in the insurance dataset.