File is added to predict the doctor's consultation fees by using regression algorithm
We have all been in situation where we go to a doctor in emergency and find that the consultation fees are too high. As a data scientist we all should do better. What if you have data that records important details about a doctor and you get to build a model to predict the doctor’s consulting fee.
File is added to predict the doctor's consultation fees by using regression algorithm